requirements.txt
flask>=2.0.0
gunicorn>=20.0.0
google-cloud-bigquery>=3.0.0
